requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pypdf==4.2.0
python-docx==1.1.0
PyYAML==6.0.2